pyproject.toml
[project]
name = "tps"
version = "0.1.0"
description = "Training Studio - GitHub-integrated Q&A analytics platform"
authors = [
    {name = "monk-ee",email = "magic.monkee.magic@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13.0a1"
dependencies = [
    "openai (>=1.98.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "faiss-cpu (>=1.11.0.post1,<2.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "ollama (>=0.5.2,<0.6.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "bert-score (>=0.3.13,<0.4.0)",
    "fastapi (>=0.104.1,<0.105.0)",
    "uvicorn (>=0.24.0,<0.25.0)", 
    "sqlalchemy (>=2.0.23,<2.1.0)",
    "alembic (>=1.13.1,<1.14.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "pydantic (>=2.5.0,<3.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "redis (>=5.0.1,<6.0.0)",
    "aiohttp (>=3.9.1,<4.0.0)",
    "python-multipart (>=0.0.6,<0.1.0)",
    "celery (>=5.3.4,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pytest = "^8.4.1"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.8.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
testpaths = ["training-studio/backend/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=training-studio/backend/app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

